Phase C: M-1 admin split + Tier-1 complexity gate skeleton
M-1 — split routes/admin.py (1239 SLOC, MI 5.12) into a package:

- routes/admin/__init__.py — assembles the unified /admin/ router
- routes/admin/_common.py — shared constants + helpers
- routes/admin/index.py — the /admin/ dashboard
- routes/admin/observability.py — read-only views
- routes/admin/actions.py — operator mutations
- routes/admin/work_refs.py — work-ref GC

URL surface unchanged (every route still mounts at /admin/...).
Each sub-module < 500 SLOC. 32 routes preserved verbatim.

Tier-1 — scripts/check-complexity.py:

- Thresholds: file SLOC > 800, file MI < 20, function CC > 20,
  function length > 100.
- # slop-allow: <justification> annotation escape hatch (per
  operator's design): function-level + file-level via
  # slop-allow-file.
- Modes: default CI fail-loud, --list report-only, --json
  machine-readable.
- Calls `radon` once over reference/ (raw + mi + cc) then walks the
  AST for function spans; no per-file subprocess fan-out.

Not wired into CI yet — that lands with the grandfathering pass that
adds operator-reviewed slop-allow annotations to the existing
violators.
